test(integration): consolidate threaded tests with rayon seed
Merge alloc_threaded and cross_thread into a single
threaded_integration.rs binary with rayon seed pre-population for
cross_thread_captures_all_calls. The rayon alloc test pins Rust 1.91.0
and correctly skips seed pre-population to avoid fingerprint mismatch.
